% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airline_sentiment_tweets}
\alias{airline_sentiment_tweets}
\title{airline_sentiment_tweets TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11541 rows and 2 columns.
}
\usage{
airline_sentiment_tweets
}
\description{
airline_sentiment_tweets TODO
}
\keyword{datasets}
