% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airports}
\alias{airports}
\title{airports TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3376 rows and 7 columns.
}
\usage{
airports
}
\description{
airports TODO
}
\keyword{datasets}
