% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{autoMPG}
\alias{autoMPG}
\alias{auto_mpg}
\title{Auto MPG Data Set}
\format{
A matrix containing 398 observations and 10 attributes.
}
\source{
http://archive.ics.uci.edu/ml/datasets/Auto+MPG
}
\usage{
auto_mpg
}
\description{
Revised from CMU StatLib library, data concerns city-cycle fuel consumption.
}
\references{
Quinlan,R. (1993). Combining Instance-Based and Model-Based
Learning. In Proceedings on the Tenth International Conference of Machine
Learning, 236-243, University of Massachusetts, Amherst. Morgan Kaufmann.
}
\keyword{datasets}
