% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbc_articles}
\alias{bbc_articles}
\alias{BBC_Articles}
\title{BBC articles data}
\format{
A data frame with 201,571 observations, each a word on a document. 
 \describe{
   \item{\code{word}}{A word in an article}
   \item{\code{document}}{The document in which that word was found}
 }
}
\source{
Unknown as of yet
}
\usage{
bbc_articles

BBC_Articles
}
\description{
The new dataset is.... and for back.....(used in book), but TV-style name used i  training ciurss
}
\keyword{datasets}
