% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbc_articles_full}
\alias{bbc_articles_full}
\title{Full BBC Articles data}
\format{
A tibble, with 927 observations of separate documents and their contents. This results in two columns.
   \describe{
      \item{\code{words}}{The words from a given article}
      \item{\code{document}}{The 'document' (article) ID}
  }
}
\source{
\itemize{\item{\url{https://www.bbc.co.uk/news}}}
}
\usage{
bbc_articles_full
}
\description{
Full BBC Articles data
}
\details{
A collection of business and politics BBC news articles. Each row represents each article (document), 
with a document ID and a string of the text content with stop words removed. This is a 'dirty' version of the 
\code{bbc_articles} dataset, where we now have a string of text for each observation, as opposed to a single word.
}
\keyword{datasets}
