% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbc_politics_123}
\alias{bbc_politics_123}
\title{bbc_politics_123 TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 86 rows and 2 columns.
}
\usage{
bbc_politics_123
}
\description{
bbc_politics_123 TODO
}
\keyword{datasets}
