% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{body_image}
\alias{body_image}
\title{body_image TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 246 rows and 8 columns.
}
\usage{
body_image
}
\description{
body_image TODO
}
\keyword{datasets}
