% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{book_sections}
\alias{book_sections}
\title{Gutenberg Project books dataset}
\format{
A tibble with 108,657 observations, each a word on a document. 
This data set is designed to show how LDA can be used to separate a set of 
mixed documents into two distinct "topics" (or books).

\describe{
 \item{\code{word}}{Words from a given section within a book.}
 \item{\code{document}}{The book section ID that the word came from.}
 }
}
\source{
Data taken from two books of the Gutenberg Project 
   \itemize{\item{\url{https://www.gutenberg.org/}}}
}
\usage{
book_sections
}
\description{
A mixed up collection of words from different book sections of two books.
}
\keyword{datasets}
