% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston}
\alias{boston}
\title{Housing Values in Suburbs of Boston
The \code{Boston} data frame has 506 rows and 14 columns.}
\format{
This data frame contains the following columns: 
\describe{
   \item{list("crim")}{Per capita crime rate by town.} 
   \item{list("zn")}{proportion of residential land zoned for lots over 25,000 sq.ft.}
   \item{list("indus")}{ proportion of non-retail business acres per town.  }
   \item{list("chas")}{ Charles River dummy variable (= 1 if tract bounds river; 0 otherwise).  } 
   \item{list("nox")}{ nitrogen oxides concentration (parts per 10 million).  } 
   \item{list("rm")}{ average number of rooms per dwelling.  } 
   \item{list("age")}{ proportion of owner-occupied units built prior to 1940.  } 
   \item{list("dis")}{ weighted mean of distances to five Boston employment centres.  } 
   \item{list("rad")}{ index of accessibility to radial highways.  } 
   \item{list("tax")}{ full-value property-tax rate per \$10,000.  } 
   \item{list("ptratio")}{ pupil-teacher ratio by town.  }
   \item{list("black")}{ \eqn{1000(Bk - 0.63)^2} where \eqn{Bk} is the proportion of blacks by town.  } 
   \item{list("lstat")}{ lower status of the population (percent).  } 
   \item{list("medv")}{ median value of owner-occupied homes in \$1000s.  }
 }
}
\source{
Harrison, D. and Rubinfeld, D.L. (1978) Hedonic prices and the
demand for clean air.  \emph{J. Environ. Economics and Management} \bold{5},
81--102.

Belsley D.A., Kuh, E.  and Welsch, R.E. (1980) \emph{Regression Diagnostics.
Identifying Influential Data and Sources of Collinearity.} New York: Wiley.
}
\usage{
boston
}
\description{
Housing Values in Suburbs of Boston
The \code{Boston} data frame has 506 rows and 14 columns.
}
\keyword{datasets}
