% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast_cancer_clean_target}
\alias{breast_cancer_clean_target}
\title{Wisconsin Breast Cancer Database}
\format{

}
\source{
\itemize{ \item Creator: Dr. WIlliam H. Wolberg (physician);
University of Wisconsin Hospital ;Madison; Wisconsin; USA \item Donor: Olvi
Mangasarian (mangasarian@cs.wisc.edu) \item Received: David W. Aha
(aha@cs.jhu.edu) } These data have been taken from the UCI Repository Of
Machine Learning Databases at \itemize{ \item
\url{ftp://ftp.ics.uci.edu/pub/machine-learning-databases} \item
\url{http://www.ics.uci.edu/~mlearn/MLRepository.html} } and were converted
to R format by Evgenia Dimitriadou.
}
\usage{
breast_cancer_clean_target
}
\description{
Wisconsin Breast Cancer Database
}
\keyword{datasets}
