% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bridges}
\alias{bridges}
\title{bridges TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 107 rows and 13 columns.
}
\usage{
bridges
}
\description{
bridges TODO
}
\keyword{datasets}
