% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carriers}
\alias{carriers}
\title{Carrier data}
\format{
A dataframe with 1492 observations and 2 variables
\describe{
   \item{\code{Code}}{A character string giving the IATA code for the carrier}
   \item{\code{Description}}{Carrier name/description}
   }
}
\usage{
carriers
}
\description{
This data comes from the RITA/Transtats database
}
\keyword{datasets}
