% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{commute}
\alias{commute}
\title{R For Data Science tidytuesday commute dataset}
\format{
A tibble containing 3,496 observations of 9 variables
   \describe{
      \item{\code{city}}{City}
      \item{\code{state}}{State}
      \item{\code{city_size}}{City Size - \itemize{\item Small = 20K to 99,999 \item Medium = 100K to 199,999 \item Large = >= 200K}}
      \item{\code{mode}}{Mode of transport, either walk or bike}
      \item{\code{n}}{Number of individuals}
      \item{\code{percent}}{Percent of total individuals}
      \item{\code{moe}}{Margin of Error (percent)}
      \item{\code{state_abb}}{Abbreviated state name}
      \item{\code{state_region}}{ACS State region}
   }
}
\source{
American Community Survey, United States Census Bureau
    \itemize{\item{R For Data Science repository: \url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2019/2019-11-05}}
             \item{Article and underlying data can be found at: \url{https://www.census.gov/library/publications/2014/acs/acs-25.html?#}}
             }
}
\usage{
commute
}
\description{
Data from the ACS Survey detailing the use of different transport modes
}
\keyword{datasets}
