% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demographics data}
\format{
A data frame with 33 observations on the following 7 demographic variables. 
This data is designed so that it can be merged with the dataset pkData. 
 \describe{
   \item{\code{Subject}}{A numeric vector giving the subject identifier}
   \item{\code{Sex}}{A factor with levels \code{F} \code{M}}
   \item{\code{Age}}{A numeric vector giving the age of the subject}
   \item{\code{Weight}}{A numeric vector giving weight in kg}
   \item{\code{Height}}{A numeric vector giving height in cm}
   \item{\code{BMI}}{A numeric vector giving the subject body mass index}
   \item{\code{Smokes}}{A factor with levels \code{No} \code{Yes}}
 }
}
\source{
Simulated data
}
\usage{
demo_data
}
\description{
Demographics data
}
\keyword{datasets}
