% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dow_jones_data}
\alias{dow_jones_data}
\alias{dowJonesData}
\title{Dow Jones Index Data}
\format{
A data frame with 252 observations on the following 7 variables containing data
from  2014-01-01 to 2015-01-01. 
 \describe{
   \item{\code{Date}}{Date of observation in character string format "\%m/\%d/\%Y" }
   \item{\code{DJI.Open}}{Opening value of DJI on the specified date}
   \item{\code{DJI.High}}{High value of the DJI on the specified date}
   \item{\code{DJI.Low}}{Low value of the DJI on the specified date}
   \item{\code{DJI.Close}}{Closing value of the DJI on the specified date}
   \item{\code{DJI.Volume}}{the number of shares or contracts traded}
   \item{\code{DJI.Adj.Close}}{Close price adjusted for dividends and splits}
 }
}
\source{
Data obtained using \code{yahooSeries} from the \code{fImport} package.
}
\usage{
dow_jones_data

dowJonesData
}
\description{
Dataset containing the Dow Jones Index between 2014-01-01 and 2015-01-01, which is a stock market index that measures the stock performance of 30 large 
companies listed on stock exchanges in the United States.
}
\details{
This dataset has be renamed using tidyverse-style snake_case naming conventions. However the original name of the dataset has been kept to ensure 
backwards compatibility with the book SAMS Teach Yourself R in 24 Hours (ISBN: 978-0-672-33848-9).
}
\keyword{datasets}
