% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugs}
\alias{drugs}
\title{Repeated Measures Drug data}
\format{
A data frame with 20 observations on the following 3 variables.
 \describe{
   \item{\code{Subj}}{A numeric vector, giving the subject ID}
   \item{\code{Drug}}{A numeric vector giving the drug ID, numbered 1 to 4}
   \item{\code{Value}}{A numeric vector, giving the observation value}
 }
}
\source{
Generated from example data used in \url{https://www.stattutorials.com/SAS/TUTORIAL-PROC-GLM-REPEAT.htm}
}
\usage{
drugs
}
\description{
Repeated Measures Drug data
}
\keyword{datasets}
