% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emax_data}
\alias{emax_data}
\alias{emaxData}
\title{Data that can be used to fit or plot Emax models}
\format{
A data frame with 64 observations on the following 6 variables.
 \describe{
   \item{\code{Subject}}{a numeric vector giving the unique subject ID}
   \item{\code{Dose}}{a numeric vector giving the dose group}
   \item{\code{E}}{a numeric vector giving the Emax}
   \item{\code{Gender}}{a numeric vector giving the gender}
   \item{\code{Age}}{a numeric vector giving the age of the subject}
   \item{\code{Weight}}{a numeric vector giving the weight}
 }
}
\source{
Simulated data
}
\usage{
emax_data

emaxData
}
\description{
This dataset has be renamed using tidyverse-style snake_case naming conventions. However the original name of the dataset has been kept to ensure 
backwards compatibility with the book SAMS Teach Yourself R in 24 Hours (ISBN: 978-0-672-33848-9).
}
\keyword{datasets}
