% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxData.R
\docType{data}
\name{emaxData}
\alias{emaxData}
\title{Data that can be used to fit or plot Emax models}
\format{A data frame with 64 observations on the following 6 variables.
 \describe{
   \item{\code{Subject}}{a numeric vector giving the unique subject ID}
   \item{\code{Dose}}{a numeric vector giving the dose group}
   \item{\code{E}}{a numeric vector giving the Emax}
   \item{\code{Gender}}{a numeric vector giving the gender}
   \item{\code{Age}}{a numeric vector giving the age of the subject}
   \item{\code{Weight}}{a numeric vector giving the weight}
 }}
\source{
Simulated data
}
\usage{
emaxData
}
\description{
Data that can be used to fit or plot Emax models
}
\keyword{datasets}

