% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxFun.R
\name{emaxFun}
\alias{emaxFun}
\title{Function to calculate Emax}
\usage{
emaxFun(Dose, E0 = 0, ED50 = 50, Emax = 100)
}
\arguments{
\item{Dose}{User provided dose values}

\item{E0}{Effect at time 0}

\item{ED50}{50\% of maximum effect}

\item{Emax}{Maximum effect}
}
\description{
Calculation used for Emax in Mango Training materials
}
\examples{
emaxFun(Dose = 100)

}

