% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global_temps}
\alias{global_temps}
\title{global_temps TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2040 rows and 3 columns.
}
\usage{
global_temps
}
\description{
global_temps TODO
}
\keyword{datasets}
