% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imdb_reviews}
\alias{imdb_reviews}
\title{imdb_reviews TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 25000 rows and 2 columns.
}
\usage{
imdb_reviews
}
\description{
imdb_reviews TODO
}
\keyword{datasets}
