% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{london}
\alias{london}
\title{london TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 33 rows and 4 columns.
}
\usage{
london
}
\description{
london TODO
}
\keyword{datasets}
