% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mangoTraining-package.R
\docType{package}
\name{mangoTraining-package}
\alias{mangoTraining-package}
\alias{mangoTraining}
\title{Mango SOlutions Training Datasets}
\description{
Datasets designed to be used in conjunction with Mango Solutions training materials.
}
\details{
Datasets designed to be used in conjunction with Mango Solutions' training materials and 
book, SAMS Teach Yourself R in 24 Hours (ISBN: 978-0-672-33848-9). The data covers a range of applications and has been collected together from a number of sources.
The airquality dataset, from the Core R datasets package is also provided in xlsx format
in the extdata directory of this package.
}
\author{
Mango Solutions

Maintainer: Aimee Gott \email{agott@mango-solutions.com}
Contact: Mango Solutions \email{rin24hours@mango-solutions.com}
}
