% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measles}
\alias{measles}
\title{measles TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3876 rows and 4 columns.
}
\usage{
measles
}
\description{
measles TODO
}
\keyword{datasets}
