% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{messy_data}
\alias{messy_data}
\alias{messyData}
\title{Messy clinical trial data}
\format{
A data frame with 33 observations on the following 7 variables. 
This data has been designed to show reshaping/tidying of data.
\describe{
  \item{\code{Subject}}{A numeric vector giving the subject ID}
  \item{\code{Placebo.1}}{A numeric vector giving the subjects observed value on treatment Placebo at time 1}
  \item{\code{Placebo.2}}{A numeric vector giving the subjects observed value on treatment Placebo at time 2}
  \item{\code{Drug1.1}}{A numeric vector giving the subjects observed value on treatment Drug1 at time 1}
  \item{\code{Drug1.2}}{A numeric vector giving the subjects observed value on treatment Drug1 at time 2}
  \item{\code{Drug2.1}}{A numeric vector giving the subjects observed value on treatment Drug2 at time 1}
  \item{\code{Drug2.2}}{A numeric vector giving the subjects observed value on treatment Drug2 at time 2}
 }
}
\source{
Simulated data
}
\usage{
messy_data

messyData
}
\description{
Simulated dataset for examples of reshaping data
}
\details{
This dataset has be renamed using tidyverse-style snake_case naming conventions. However the original name of the dataset has been kept to ensure 
backwards compatibility with the book SAMS Teach Yourself R in 24 Hours (ISBN: 978-0-672-33848-9).
}
\keyword{datasets}
