% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingPk.R
\docType{data}
\name{missingPk}
\alias{missingPk}
\title{Clinical trial data}
\format{A data frame with 165 observations on the following 4 variables.
\describe{
  \item{\code{Subject}}{a numeric vector}
  \item{\code{Dose}}{a numeric vector}
  \item{\code{Time}}{a numeric vector}
 \item{\code{Conc}}{a numeric vector}
 }}
\source{
Simulated from pkData
}
\usage{
missingPk
}
\description{
Clinical trial data
}
\keyword{datasets}

