% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{olympics}
\alias{olympics}
\title{olympics TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 454 rows and 9 columns.
}
\usage{
olympics
}
\description{
olympics TODO
}
\keyword{datasets}
