% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pk_data}
\alias{pk_data}
\alias{pkData}
\title{Typical PK data}
\format{
A data frame with 165 observations on the following 4 variables.
 \describe{
   \item{\code{Subject}}{a numeric vector giving the subject identifier}
   \item{\code{Dose}}{a numeric vector giving the dose group}
   \item{\code{Time}}{a numeric vector giving the observation times}
   \item{\code{Conc}}{a numeric vector giving the observed concentration}
 }
}
\source{
Simulated data
}
\usage{
pk_data

pkData
}
\description{
Typical PK data
}
\keyword{datasets}
