% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkData.R
\docType{data}
\name{pkData}
\alias{pkData}
\title{Typical PK data}
\format{A data frame with 165 observations on the following 4 variables.
 \describe{
   \item{\code{Subject}}{a numeric vector}
   \item{\code{Dose}}{a numeric vector}
   \item{\code{Time}}{a numeric vector}
   \item{\code{Conc}}{a numeric vector}
 }}
\source{
Simulated data
}
\usage{
pkData
}
\description{
Typical PK data
}
\keyword{datasets}

