% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{policy_data}
\alias{policy_data}
\alias{policyData}
\title{Insurance Policy Data}
\format{
A data frame with 926 observations on the following 13 variables.
\describe{
  \item{\code{Year}}{The four digit year of the policy}
  \item{\code{PolicyNo}}{The policy number}
  \item{\code{TotalPremium}}{The total incurance premium}
 \item{\code{BonusMalus}}{Discount level}
  \item{\code{WeightClass}}{The weight class of the car}
  \item{\code{Region}}{Region of the car owner}
  \item{\code{Age}}{Age of the main driver}
 \item{\code{Mileage}}{Estimted annual mileage}
  \item{\code{Usage}}{Car usage}
  \item{\code{PremiumClass}}{Class of the car}
  \item{\code{NoClaims}}{Number of previous claims}
 \item{\code{GrossIncurred}}{Claim cost}
 \item{\code{Exposure}}{How long they have been driving}
 }
}
\source{
Simulated based on details of how to simulate car insurance data in
Modern Actuarial Risk Theory Using R 2nd Edition (Rob Kaas, Marc Goovaerts, Jan Dhaene, Michel Denuit)
}
\usage{
policy_data

policyData
}
\description{
Insurance Policy Data
}
\details{
This dataset has be renamed using tidyverse-style snake_case naming conventions. However the original name of the dataset has been kept to ensure 
backwards compatibility with the book SAMS Teach Yourself R in 24 Hours (ISBN: 978-0-672-33848-9).
}
\keyword{datasets}
