% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtpk2.R
\docType{data}
\name{qtpk2}
\alias{qtpk2}
\title{Typical PK data}
\format{A data frame with 2061 observations on the following 8 variables.
 \describe{
   \item{\code{subjid}}{A numeric vector giving the subject ID}
   \item{\code{treat}}{A factor giving the treatment}
   \item{\code{time}}{A numeric vector giving the observation times}
   \item{\code{qt}}{A numeric vector giving the QT interval value}
   \item{\code{qtcb}}{A numeric vector giving corrected QT interval}
   \item{\code{hr}}{A numeric vector giving the heart rate}
   \item{\code{rr}}{A numeric vector giving the R-R interval}
   \item{\code{sex}}{A factor giving the subject sex}
 }}
\source{
A subset of the data qtpk originally provided in the QT package
}
\usage{
qtpk2
}
\description{
Typical PK data
}
\keyword{datasets}

