% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{run_data}
\alias{run_data}
\alias{runData}
\title{An example of NONMEM run data}
\format{
A data frame with 73 observations on the following 10 variables.
\describe{
   \item{\code{ID}}{a numeric vector giving the subject ID}
   \item{\code{DAY}}{a numeric vector giving the day of the observation}
   \item{\code{CL}}{a numeric vector giving the clearance value}
   \item{\code{V}}{a numeric vector giving the volume of distribution}
   \item{\code{WT}}{a numeric vector giving the weight}
   \item{\code{DV}}{a numeric vector giving the dependent variable}
   \item{\code{IPRE}}{a numeric vector giving the individual prediction}
   \item{\code{PRED}}{a numeric vector giving the population prediction}
   \item{\code{RES}}{a numeric vector giving the residual}
   \item{\code{WRES}}{a numeric vector giving the weighted residual}
 }
}
\source{
Simulated Data
}
\usage{
run_data

runData
}
\description{
This dataset has be renamed using tidyverse-style snake_case naming conventions. However the original name of the dataset has been kept to ensure 
backwards compatibility with the book SAMS Teach Yourself R in 24 Hours (ISBN: 978-0-672-33848-9).
}
\keyword{datasets}
