% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runData.R
\docType{data}
\name{runData}
\alias{runData}
\title{An example of NONMEM run data}
\format{A data frame with 73 observations on the following 10 variables.
\describe{
   \item{\code{ID}}{a numeric vector}
   \item{\code{DAY}}{a numeric vector}
   \item{\code{CL}}{a numeric vector}
   \item{\code{V}}{a numeric vector}
   \item{\code{WT}}{a numeric vector}
   \item{\code{DV}}{a numeric vector}
   \item{\code{IPRE}}{a numeric vector}
   \item{\code{PRED}}{a numeric vector}
   \item{\code{RES}}{a numeric vector}
   \item{\code{WRES}}{a numeric vector}
 }}
\source{
Simulated Data
}
\usage{
runData
}
\description{
An example of NONMEM run data
}
\keyword{datasets}

