% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students}
\alias{students}
\title{students TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 146 rows and 15 columns.
}
\usage{
students
}
\description{
students TODO
}
\keyword{datasets}
