% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{word_weights}
\alias{word_weights}
\title{word_weights TODO}
\format{
An object of class \code{matrix} with 61327 rows and 100 columns.
}
\usage{
word_weights
}
\description{
word_weights TODO
}
\keyword{datasets}
