% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xp_data}
\alias{xp_data}
\alias{xpData}
\title{Typical NONMEM data}
\format{
A data frame with 1061 observations on the following 23 variables.
\describe{
   \item{\code{ID}}{a numeric vector giving the subject ID}
   \item{\code{SEX}}{a numeric vector giving the subject sex}
   \item{\code{RACE}}{a numeric vector giving the subject race}
   \item{\code{SMOK}}{a numeric vector giving the subject smoking status}
   \item{\code{HCTZ}}{a numeric vector giving the treatment status}
   \item{\code{PROP}}{a numeric vector giving the treatment status}
   \item{\code{CON}}{a numeric vector giving the treatment status}
   \item{\code{DV}}{a numeric vector giving the dependent variable}
   \item{\code{PRED}}{a numeric vector giving population prediction}
   \item{\code{RES}}{a numeric vector giving the residual}
   \item{\code{WRES}}{a numeric vector giving the weighted residual}
   \item{\code{AGE}}{a numeric vector giving the subject age}
   \item{\code{HT}}{a numeric vector giving the subject height}
   \item{\code{WT}}{a numeric vector giving the subject weight}
   \item{\code{SECR}}{a numeric vector giving the serum creatinine value}
   \item{\code{OCC}}{a numeric vector giving the occasion}
   \item{\code{TIME}}{a numeric vector giving the time of the observation time}
   \item{\code{IPRE}}{a numeric vector giving individual prediction}
   \item{\code{IWRE}}{a numeric vector giving the individual weighted residual}
   \item{\code{SID}}{a numeric vector giving the site ID}
   \item{\code{CL}}{a numeric vector giving the clearance}
   \item{\code{V}}{a numeric vector giving the volume of distribution}
   \item{\code{KA}}{a numeric vector giving the absorption rate constant}
 }
}
\source{
Simulated Data
}
\usage{
xp_data

xpData
}
\description{
This dataset has be renamed using tidyverse-style snake_case naming conventions. However the original name of the dataset has been kept to ensure 
backwards compatibility with the book SAMS Teach Yourself R in 24 Hours (ISBN: 978-0-672-33848-9).
}
\keyword{datasets}
