% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbc_articles.R
\docType{data}
\name{bbc_articles}
\alias{bbc_articles}
\title{BBC articles data}
\format{
A data frame with 201,571 observations, each a word on a document. 
 \describe{
   \item{\code{word}}{A word in an article}
   \item{\code{document}}{The document in which that word was found}
 }
}
\source{
Unknown as of yet
}
\usage{
bbc_articles
}
\description{
BBC articles data
}
\keyword{datasets}
