% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbc_articles}
\alias{bbc_articles}
\title{BBC articles data}
\format{
A tibble with 201,571 observations, each a word on a document.
 \describe{
   \item{\code{word}}{A word in an article}
   \item{\code{document}}{The document/article ID where the word was taken from}
 }
}
\source{
\itemize{\item{\url{https://www.bbc.co.uk/news}}}
}
\usage{
bbc_articles
}
\description{
A collection of BBC news articles from the business or politics sections. There are a total of 927 articles used.
}
\keyword{datasets}
