% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbc_politics_123}
\alias{bbc_politics_123}
\title{BBC Politics article data}
\format{
A tibble with 86 observations, each a word on a document.
 \describe{
   \item{\code{word}}{A word in an article}
   \item{\code{document}}{The document/article ID where the word was taken from. Note: this only has one unique value, 
   however the column is kept for comparison with other BBC datasets.}
 }
}
\source{
\itemize{\item{\url{https://www.bbc.co.uk/news}}}
}
\usage{
bbc_politics_123
}
\description{
A single BBC Politics article (not included in the full BBC articles dataset), given in tidy, one word per row format.
}
\keyword{datasets}
