% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{body_image}
\alias{body_image}
\title{Body image dataset}
\format{
A tibble of 246 observations on 8 attributes.
\describe{
   \item{\code{ethnicity}}{Subject's ethnicity (Asian, Europn, Maori, Pacific)}
   \item{\code{married}}{How many times have they been married?}
   \item{\code{bodyim}}{Subject's rating of themselves (slight.uw, right, slight.ow, mod.ow, very.ow)}
   \item{\code{sm.ever}}{Have they ever smoked?}
   \item{\code{weight}}{Weight in kilograms}
   \item{\code{height}}{Height in centimetres}
   \item{\code{age}}{Age in years}
   \item{\code{stressgp}}{What stress group are they in?}
   }
}
\source{
Simulated data
}
\usage{
body_image
}
\description{
Body image dataset
}
\details{
A simulated dataset containing data on the self-image of subjects with differing body aesthetics
}
\keyword{datasets}
