% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{book_sections}
\alias{book_sections}
\title{book_sections TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 108657 rows and 2 columns.
}
\usage{
book_sections
}
\description{
book_sections TODO
}
\keyword{datasets}
