% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston}
\alias{boston}
\title{Boston housing dataset}
\format{
This data frame contains the following columns: 
\describe{
   \item{\code{tract}}{Census tract}
   \item{\code{medv}}{Median value of owner-occupied homes in \$1000s.  }
   \item{\code{crim}}{Per capita crime rate by town.} 
   \item{\code{zn}}{Proportion of residential land zoned for lots over 25,000 sq.ft.}
   \item{\code{indus}}{Proportion of non-retail business acres per town.  }
   \item{\code{chas}}{Charles River dummy variable (= 1 if tract bounds river; 0 otherwise).  } 
   \item{\code{nox}}{Nitrogen oxides concentration (parts per 10 million).  } 
   \item{\code{rm}}{Average number of rooms per dwelling.  } 
   \item{\code{age}}{Proportion of owner-occupied units built prior to 1940.  } 
   \item{\code{dis}}{Weighted mean of distances to five Boston employment centres.  } 
   \item{\code{rad}}{Index of accessibility to radial highways.  } 
   \item{\code{tax}}{Full-value property-tax rate per \$10,000.  } 
   \item{\code{ptratio}}{Pupil-teacher ratio by town.  }
   \item{\code{b}}{\eqn{1000(Bk - 0.63)^2} where \eqn{Bk} is the proportion of blacks by town.  } 
   \item{\code{lstat}}{Lower status of the population (percent).  } 
 }
}
\source{
Harrison, D. and Rubinfeld, D.L. (1978) Hedonic prices and the
demand for clean air.  \emph{J. Environ. Economics and Management} \bold{5},
81--102.

Belsley D.A., Kuh, E.  and Welsch, R.E. (1980) \emph{Regression Diagnostics.
Identifying Influential Data and Sources of Collinearity.} New York: Wiley.
}
\usage{
boston
}
\description{
Housing Values in Suburbs of Boston.
}
\details{
The \code{boston} data frame has 506 rows and 15 columns.
}
\keyword{datasets}
