% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{commute}
\alias{commute}
\title{commute TODO}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3496 rows and 9 columns.
}
\usage{
commute
}
\description{
commute TODO
}
\keyword{datasets}
