% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\alias{demoData}
\title{Demographics data}
\format{
A data frame with 33 observations on the following 7 demographic variables. 
This data is designed so that it can be merged with the dataset pkData. 
 \describe{
   \item{\code{Subject}}{A numeric vector giving the subject identifier}
   \item{\code{Sex}}{A factor with levels \code{F} \code{M}}
   \item{\code{Age}}{A numeric vector giving the age of the subject}
   \item{\code{Weight}}{A numeric vector giving weight in kg}
   \item{\code{Height}}{A numeric vector giving height in cm}
   \item{\code{BMI}}{A numeric vector giving the subject body mass index}
   \item{\code{Smokes}}{A factor with levels \code{No} \code{Yes}}
 }
}
\source{
Simulated data
}
\usage{
demo_data

demoData
}
\description{
This dataset has be renamed using tidyverse-style snake_case naming conventions. However the original name of the dataset has been kept to ensure 
backwards compatibility with the book SAMS Teach Yourself R in 24 Hours (ISBN: 978-0-672-33848-9).
}
\keyword{datasets}
