% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emaxFun.R
\name{emaxFun}
\alias{emaxFun}
\title{Function to calculate Emax}
\usage{
emaxFun(Dose, E0 = 0, ED50 = 50, Emax = 100)
}
\arguments{
\item{Dose}{The dose value}

\item{E0}{Response effect when Dose is 0}

\item{ED50}{Median effective dose}

\item{Emax}{Maximum effective dose}
}
\description{
Calculation used for Emax in Mango Training materials
}
\examples{
emaxFun(Dose = 100)

}

