% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_fun.R
\name{logistic_fun}
\alias{logistic_fun}
\title{Function to fit logistic model}
\usage{
logistic_fun(Dose, E0 = 0, EC50 = 50, Emax = 1, rc = 5)
}
\arguments{
\item{Dose}{The dose value to calculate at}

\item{E0}{Effect at time 0}

\item{EC50}{50\% of maximum effect}

\item{Emax}{Maximum effect}

\item{rc}{rate constant}
}
\description{
Simple logistic function as used in Mango training materials
}
\examples{
logisticFun(Dose = 50)

}
