% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticFun.R
\name{logisticFun}
\alias{logisticFun}
\title{Function to fit logistic model}
\usage{
logisticFun(Dose, E0 = 0, EC50 = 50, Emax = 1, rc = 5)
}
\arguments{
\item{Dose}{The dose value}

\item{E0}{Response effect when Dose is 0}

\item{EC50}{Half maximal effective concentration}

\item{Emax}{Maximum effective dose}

\item{rc}{rate constant}
}
\description{
Simple logistic function as used in Mango training materials
}
\examples{
logisticFun(Dose = 50)

}

