% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mangoTraining-package.R
\docType{package}
\name{mangoTraining-package}
\alias{mangoTraining}
\alias{mangoTraining-package}
\title{Mango SOlutions Training Datasets}
\description{
Datasets designed to be used in conjunction with Mango Solutions training materials.
}
\details{
Datasets designed to be used in conjunction with Mango Solutions' training materials.
The data covers a range of applications.
}
\author{
Mango Solutions

Maintainer: Aimee Gott \email{agott@mango-solutions.com}
Contact: Mango Solutions \email{support@mango-solutions.com}
}

