% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_pk}
\alias{missing_pk}
\alias{missingPk}
\title{Clinical trial data}
\format{
A data frame with 165 observations on the following 4 variables.
\describe{
  \item{\code{Subject}}{a numeric vector giving the subject identifier}
  \item{\code{Dose}}{a numeric vector giving the dose group}
  \item{\code{Time}}{a numeric vector giving the obserrvation times}
 \item{\code{Conc}}{a numeric vector giving the observed concentration}
 }
}
\source{
Simulated from pkData
}
\usage{
missing_pk

missingPk
}
\description{
This dataset has be renamed using tidyverse-style snake_case naming conventions. However the original name of the dataset has been kept to ensure 
backwards compatibility with the book SAMS Teach Yourself R in 24 Hours (ISBN: 978-0-672-33848-9).
}
\keyword{datasets}
