% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students}
\alias{students}
\title{Students simulated data}
\format{
A tibble with 146 observations of 15 variables.
\describe{
   \item{\code{Grade}}{Final grade (A, B, C, D)}
   \item{\code{Pass}}{Did they pass the course? (No, Yes)}
   \item{\code{Exam}}{Mark in final exam (out of 100)}
   \item{\code{Degree}}{The degree type undertaken by the student}
   \item{\code{Gender}}{Gender of the student}
   \item{\code{Attend}}{Did they regularly attend class? (No, Yes)}
   \item{\code{Assign}}{Score obtained in mid-term assignment (out of 20)}
   \item{\code{Test}}{Score obtained in previous term test (out of 20)}
   \item{\code{B}}{Mark for short answer section (out of 20)}
   \item{\code{C}}{Mark for long answer section (out of 20)}
   \item{\code{MC}}{Mark for multiple choice sectionC (out of 30)}
   \item{\code{Colour}}{Colour of exam booklet (Blue, Green, Pink, Yellow)}
   \item{\code{Stage1}}{Stage one grade (A, B, C)}
   \item{\code{Years.Since}}{Number of years since doing Stage 1}
   \item{\code{Repeat}}{Where they repeating the paper? (No, Yes)}
   }
}
\source{
Simulated data
}
\usage{
students
}
\description{
Students simulated data
}
\keyword{datasets}
